# module_02_03/AI_system_interview_design/openapi.yaml
openapi: 3.1.0
info:
  title: Loopboard API
  version: 0.1.0
  summary: Backend contract expected by the Loopboard frontend (system design interview platform).
  description: |
    This document specifies the HTTP backend that the existing frontend expects. It was derived
    from the frontend's API client, `frontend/src/lib/mock-api.ts`, which is a localStorage +
    BroadcastChannel stand-in for a real server. Every exported function in that module maps to
    exactly one operation below, and every payload shape is taken from
    `frontend/src/lib/canvas-types.ts` and the type declarations in `mock-api.ts`.

    ## Domain

    An **interviewer** creates a *session* (a shared architecture canvas), shares a join link,
    and one or more **candidates** / **observers** join it by opening the link and entering a
    display name. Everyone edits the same canvas document and sees each other's cursors live.
    Ending a session freezes the canvas read-only; it can be reopened.

    ## Behaviour the frontend relies on

    * **Timestamps** (`createdAt`, `endedAt`, `lastSeen`, `at`) are epoch **milliseconds**, not
      seconds — the frontend passes them straight to `new Date(...)` and compares them to
      `Date.now()`.
    * **`GET /sessions` must be sorted** by `createdAt` descending; the dashboard renders the
      array as-is.
    * **Cursors older than 15 s must be omitted** from `GET /sessions/{sessionId}/cursors`. The
...
      `"Guest"`.
...

    ## Realtime

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - participantAuth: []
...
paths:
...
  /sessions/{sessionId}/canvas:
    parameters:
      - $ref: "#/components/parameters/SessionId"

    get:
      tags: [Canvas]
      operationId: getCanvas
      summary: Read the whole canvas document
      description: |
        Returns the full node list. Called on mount and after every `doc` event — this is how late
        joiners and reconnecting clients resync (spec §5). An empty board is `{"nodes": []}`, never
        `404`, as long as the session exists.
      responses:
        "200":
          description: The canvas document.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CanvasDoc"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [Canvas]
      operationId: saveCanvas
      summary: Replace the whole canvas document
      description: |
        Last-write-wins whole-document replace — this is what the frontend actually calls today,
        on a 120 ms debounce after any edit (`handleChange` in `session.$sessionId.tsx`). It is
        the only canvas write path currently in use; the node-level operations below exist in the
        client module but are not yet wired to the UI.

        A whole-document PUT under concurrent editing will drop edits — two people dragging at
        once means the slower client's document overwrites the faster one's. The mock's own
        comment flags this ("Real backend would send CRDT ops instead"), and spec §5 calls for a
        CRDT. Treat this operation as the MVP contract and plan to add a CRDT/op-based channel
        beside it, not as the end state.

        Rejected with `409` when the session is ended or `canvasLocked`, and `403` for observers.
        Emits a `doc` event to every subscriber **except**, ideally, the writer.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CanvasDoc"
      responses:
        "204":
          description: Saved.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/CanvasLocked"
        "413":
          description: Document exceeds the server's size limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /sessions/{sessionId}/canvas/nodes:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    patch:
      tags: [Canvas]
      operationId: upsertNodes
      summary: Insert or update individual nodes
      description: |
        Merges the given nodes into the document by `id` — existing ids are replaced, new ids
        appended. Nodes not mentioned are untouched. Same lock/role rules and `doc` event as
        `PUT /canvas`.

        Node ids are generated **client-side** (`newNodeId()` → `n_` + 10 chars), so the server
        must accept ids it has never seen and must not renumber them.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [nodes]
              properties:
                nodes:
                  type: array
                  items:
                    $ref: "#/components/schemas/CanvasNode"
      responses:
        "204":
          description: Merged.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/CanvasLocked"
  /sessions/{sessionId}/canvas/nodes/delete:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    post:
      tags: [Canvas]
      operationId: deleteNodes
      summary: Delete nodes by id
      description: |
        Removes every node whose id appears in `ids`; unknown ids are ignored. Modelled as `POST`
        rather than `DELETE` because the id list is a body, and multi-select deletion can send
        many ids at once. Same lock/role rules and `doc` event as `PUT /canvas`.

        Note that deleting a shape does **not** delete arrows anchored to it — the frontend
        renders a `{shapeId}` endpoint whose shape is gone by falling back to the arrow's last
        known point. Cascading is a product decision, not a requirement of this contract.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [ids]
              properties:
                ids:
                  type: array
                  items:
                    type: string
                  examples:
                    - ["n_4k2mq7xr9b", "n_pt3wz8dnv2"]
      responses:
        "204":
          description: Deleted.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/CanvasLocked"
...
components:
  securitySchemes:
    participantAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Session-scoped token issued by `POST /sessions/{sessionId}/participants`. Encodes the
        participant id, the session id and the role. Must be rejected for any other session's
        endpoints. Should expire with the session (plus a grace period for reviewing ended
        boards).

    interviewerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Account-scoped token for a signed-in interviewer (spec §2/§6: email+password or SSO).
        Required by the dashboard operations that are not addressed by a known session id.
        **Not yet obtained by the frontend** — there is no sign-in screen; see the "Gap to close"
        note in the API description.
  parameters:
    SessionId:
      name: sessionId
      in: path
      required: true
      description: |
        Unguessable session identifier. The mock generates 16 characters from the
        ambiguity-free alphabet `abcdefghijkmnpqrstuvwxyz23456789`; the server owns generation and
        may use any format, but it must stay URL-safe (it appears in the join link) and long
        enough to be unguessable (spec §7).
      schema:
        type: string
        minLength: 8
        maxLength: 64
      examples:
        default:
          value: k7m2xp9qrt4wnb3d
...
  responses:
    BadRequest:
      description: Malformed body or failed validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing, malformed or expired token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: |
        Authenticated but not permitted — an observer attempting a canvas write, a non-host
        attempting to end/reopen/delete a session or remove another participant, or a token scoped
        to a different session or participant.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: |
        No such session (or it was deleted). The frontend treats this as "this link isn't valid"
        and must receive a `404` rather than a `200` with an empty payload.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    CanvasLocked:
      description: The session is ended or the canvas is locked, so writes are rejected.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    ComponentKind:
      type: string
      description: |
        The fixed component palette from `canvas-types.ts` (spec §4.1). Curated and closed for the
        MVP — custom icons are Phase 2 — so the server may validate against this enum. Each kind
        carries a default geometry and size in the frontend; those are presentation concerns and
        are not part of the wire format.
      enum:
        - client
        - load-balancer
        - api-gateway
        - service
        - database
        - nosql
        - cache
        - queue
        - llm
        - cdn
        - storage
        - box
    Endpoint:
      description: |
        One end of an arrow: either anchored to a shape (the arrow follows it when the shape
        moves, spec §4.2) or pinned to a free canvas point.
      oneOf:
        - type: object
          additionalProperties: false
          required: [shapeId]
          properties:
            shapeId:
              type: string
              description: Id of a `ShapeNode` in the same document.
        - type: object
          additionalProperties: false
          required: [x, y]
          properties:
            x:
              type: number
            y:
              type: number
    ShapeNode:
      type: object
      additionalProperties: false
      required: [type, id, kind, label, x, y, w, h, authorId]
      properties:
        type:
          const: shape
        id:
          type: string
          examples: ["n_4k2mq7xr9b"]
        kind:
          $ref: "#/components/schemas/ComponentKind"
        label:
          type: string
          description: Editable — the candidate renames a generic `service` to e.g. "Auth Service".
        x:
          type: number
        y:
          type: number
        w:
          type: number
        h:
          type: number
        authorId:
          type: string
          description: The `Participant.id` that created the node.
    ArrowNode:
      type: object
      additionalProperties: false
      required: [type, id, from, to, label, dashed, bidirectional, authorId]
      properties:
        type:
          const: arrow
        id:
          type: string
        from:
          $ref: "#/components/schemas/Endpoint"
        to:
          $ref: "#/components/schemas/Endpoint"
        label:
          type: string
          description: e.g. "writes to", "async", "1:N". Empty string when unlabelled.
        dashed:
          type: boolean
          description: Dashed conventionally means an async call.
        bidirectional:
          type: boolean
        authorId:
          type: string
    StrokeNode:
      type: object
      additionalProperties: false
      required: [type, id, points, color, width, authorId]
      properties:
        type:
          const: stroke
        id:
          type: string
        points:
          type: array
          description: |
            Freehand path as an ordered list of `[x, y]` canvas-space pairs, appended to as the
            pointer moves. A long stroke can hold hundreds of pairs — the dominant contributor to
            document size.
          items:
            type: array
            items:
              type: number
            minItems: 2
            maxItems: 2
          examples:
            - [[120, 340], [122, 344], [127, 351]]
        color:
          type: string
        width:
          type: number
        authorId:
          type: string
    TextNode:
      type: object
      additionalProperties: false
      required: [type, id, x, y, text, authorId]
      properties:
        type:
          const: text
        id:
          type: string
        x:
          type: number
        y:
          type: number
        text:
          type: string
          description: Created empty and filled in place, so an empty string is valid.
        authorId:
          type: string
    CanvasNode:
      description: |
        A single element on the board, discriminated by `type`. Shapes, arrows, strokes and text
        share one flat list and one z-order (spec §4.3), so the array order is meaningful — later
        entries render on top — and the server must preserve it.
      oneOf:
        - $ref: "#/components/schemas/ShapeNode"
        - $ref: "#/components/schemas/ArrowNode"
        - $ref: "#/components/schemas/StrokeNode"
        - $ref: "#/components/schemas/TextNode"
      discriminator:
        propertyName: type
        mapping:
          shape: "#/components/schemas/ShapeNode"
          arrow: "#/components/schemas/ArrowNode"
          stroke: "#/components/schemas/StrokeNode"
          text: "#/components/schemas/TextNode"
    CanvasDoc:
      type: object
      additionalProperties: false
      required: [nodes]
      description: 'The whole board. An empty board is `{"nodes": []}`.'
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/CanvasNode"
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
          description: Stable machine-readable code.
          examples: ["session_not_found", "canvas_locked", "forbidden_role"]
        message:
          type: string
          description: Human-readable detail, safe to surface in the UI.